+++ Resources/app/administration/src/module/co-plugins/index.ts
@@ -6,12 +6,12 @@ const PluginsModule: CompanyOSModule = {
     {
       path: '/plugins',
       name: 'plugins',
-      component: () => import('./page/co-plugins-list')
+      component: () => import('./page/co-plugins-list.ts')
     }
   ],
   components: {
-    'co-plugins-grid': () => import('./component/co-plugins-grid'),
-    'co-plugin-card': () => import('./component/co-plugin-card')
+    'co-plugins-grid': () => import('./component/co-plugins-grid/index'),
+    'co-plugin-card': () => import('./component/co-plugin-card/index')
   },
   services: {
     'pluginService': () => import('./service/plugin-service')